license: allow QEMU to reuse acpi bits under GPLv2+
You are getting this mail because you might have contributed code to one
of the files in seabios that we want to reuse in QEMU,
when this file was under GPLv3 or LGPLv3.

QEMU is GPLv2 at the moment, so as a step in the process of moving acpi
tables to qemu, we need to make sure the code we'll be moving is GPLv2
compatible.

The code was originally LGPLv2 in bochs so these bits are OK.

QEMU generally prefers GPLv2 or later, so this is what this
patch does. The plan is therefore:
- collect acks from everyone
- copy code to QEMU and apply this patch to QEMU copy only

If you allow the use of your contribution in QEMU under the
terms of GPLv2 or later as proposed by this patch,
please respond to this mail including the line:

	Acked-by: Name <email address>

in the message body.

Thanks!

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>